requirements.txt
# GUI Framework
PySide6>=6.6.0

# PDF Processing
PyMuPDF>=1.23.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# Data Handling
numpy>=1.24.0

# Build
pyinstaller>=6.0.0

